As better alternative to common_py3, make common.py import on Python 3.
